[REFACTOR] 랭킹 관련 DTO,VO 위치 변경
