+++ src/main/java/com/example/demo/service/MemberService.java
@@ -20,7 +20,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.example.demo.dto.member.MemberAllMyPostsResponseDto;
 import com.example.demo.dto.member.MemberBattleResponseDto;
 import com.example.demo.dto.member.MemberBattlesResponseDto;
-import com.example.demo.dto.member.RankersResponseDto;
+import com.example.demo.dto.ranking.RankersResponseDto;
 import com.example.demo.exception.ServerNotActiveException;
 import com.example.demo.model.battle.Battle;
 import com.example.demo.model.battle.BattleStatus;
